requirements.txt
requests>=2.24,<2.25
python-dateutil>=2.8,<2.9
progressbar2>=3.53,<3.54